:recycle: 알림 부분 객체로 역할 위임
